chore: migrate offix-offline unit tests to jest
