--- packages/offix-offline/test/mocha.opts
@@ -1,3 +1,0 @@
---require ts-node/register
---require source-map-support/register
-test/**/**.ts
